add new language(tly), make Chinese work and recompile
Strangly enough gettext expects the locale for Simplified Chinese
to be called zh-cn but the directory needs to be called zh_CN.
Note further that the version of django-localeurl that is shipped
with Debian is case-sensitive with respect to locales names. This
has been fixed only in the latest development version.

@Nikerabbit This commit adds a symlinked directory locale/zh_CN
which TranslateWiki probably wants to explicitly ignore.